theme toggle moved to lib/components instead of lib/layout
