+++ src/lib/layout/Header.tsx
@@ -1,6 +1,6 @@
 import { Box, Flex } from "@chakra-ui/react";
 
-import ThemeToggle from "./ThemeToggle";
+import ThemeToggle from "../components/themeToggle/ThemeToggle";
 
 const Header = () => {
   return (
